requirements.txt
nltk
sense2vec
sentence_transformers
stanza
openai
scipy
datasets
tqdm
termcolor
pytorch-lightning